Do not use Reader and Writer classes for writing response, because it will use the default content encoding of the JRE, causing the reply from the mirror to not be an exact copy of the incoming request. Therefore, rather use InputStream and OutputStream.
If the mirror gets a header specifying the content length, we know how much data to read, and therefore we can allow the reading to block waiting for more data to appear.
For chunked transfer, we still have to only read until there is no more data, and not block to wait for more.

Add unit test for the mirror thread. This will expose the problems of the previous version of the HttpMirrorThread, and show that the version of HttpMirrorThread in this changelist works.

git-svn-id: https://svn.apache.org/repos/asf/jakarta/jmeter/trunk@584508 13f79535-47bb-0310-9956-ffa450edef68

Former-commit-id: 644bb73f9c6f6417d8d11f249dfd0cf1b488eec1